# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Shelter Center API
  version: 1.0.0
  description: "유기동물 보호소 운영 API (Animals/Adoptions/Donations/Medical/Volunteers/Stats)"
servers:
  - url: http://localhost:8080
    description: local
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    ErrorResponse:
      type: object
      required: [timestamp, path, status, code, message]
      properties:
        timestamp: { type: string, format: date-time }
        path: { type: string }
        status: { type: integer }
        code: { type: string }
        message: { type: string }
        details: { type: object, nullable: true }
    PageResponse:
      type: object
      required: [content, page, size, totalElements, totalPages, sort]
      properties:
        content:
          type: array
          items: {}
        page: { type: integer }
        size: { type: integer }
        totalElements: { type: integer }
        totalPages: { type: integer }
        sort: { type: string }
    Animal:
      type: object
      properties:
        id: { type: integer }
        name: { type: string, nullable: true }
        species: { type: string, example: dog }
        breed: { type: string, nullable: true }
        sex:
          type: string
          enum: [M, F, UNKNOWN]
        ageYears: { type: integer, nullable: true }
        intakeDate: { type: string, format: date-time }
        neutered: { type: boolean }
        status:
          type: string
          enum: [SHELTERED, TEMP_FOSTER, ADOPTED, DECEASED]
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
...
paths:
...
  /api/animals:
    post:
      tags: [Animals]
      summary: Create animal
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [species, intakeDate]
              properties:
                name: { type: string, nullable: true, maxLength: 50 }
                species: { type: string, maxLength: 20 }
                breed: { type: string, nullable: true, maxLength: 50 }
                sex: { type: string, enum: [M, F, UNKNOWN], default: UNKNOWN }
                ageYears: { type: integer, nullable: true, minimum: 0, maximum: 30 }
                intakeDate: { type: string, format: date-time }
                neutered: { type: boolean, default: false }
                status:
                  type: string
                  enum: [SHELTERED, TEMP_FOSTER, ADOPTED, DECEASED]
                  default: SHELTERED
      responses:
        "201":
          description: created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Animal" }
        "422":
          description: validation failed
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }

    get:
      tags: [Animals]
      summary: List animals (filter/search/sort/page)
      security: [{ bearerAuth: [] }]
      parameters:
        - in: query
          name: species
          schema: { type: string }
        - in: query
          name: neutered
          schema: { type: boolean }
        - in: query
          name: status
          schema:
            type: string
            enum: [SHELTERED, TEMP_FOSTER, ADOPTED, DECEASED]
        - in: query
          name: keyword
          schema: { type: string }
        - in: query
          name: from
          schema: { type: string, format: date-time }
        - in: query
          name: to
          schema: { type: string, format: date-time }
        - in: query
          name: page
          schema: { type: integer, default: 0, minimum: 0 }
        - in: query
          name: size
          schema: { type: integer, default: 20, minimum: 1, maximum: 50 }
        - in: query
          name: sort
          schema: { type: string, default: "createdAt,DESC" }
      responses:
        "200":
          description: paged animals
          content:
            application/json:
              schema:
                allOf: [{ $ref: "#/components/schemas/PageResponse" }]
                properties:
                  content:
                    type: array
                    items: { $ref: "#/components/schemas/Animal" }
        "401":
          description: Unauthorized
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
        "400":
          description: Invalid query
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
  /api/animals/{id}:
    get:
      tags: [Animals]
      summary: Get animal by id
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: animal
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Animal" }
        "404":
          description: not found
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
        "400":
          description: invalid id
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }

    patch:
      tags: [Animals]
      summary: Update animal (partial)
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              description: "validateUpdate(req.body) 기준"
      responses:
        "200":
          description: updated
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Animal" }
        "422":
          description: validation failed
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
        "404":
          description: not found
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }

    delete:
      tags: [Animals]
      summary: Delete animal
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        "204":
          description: deleted
        "404":
          description: not found
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
  /api/animals/{id}/status:
    patch:
      tags: [Animals]
      summary: Update animal status (ADMIN planned)
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [status]
              properties:
                status:
                  type: string
                  enum: [SHELTERED, TEMP_FOSTER, ADOPTED, DECEASED]
      responses:
        "200":
          description: updated
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Animal" }
        "422":
          description: validation failed
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
        "404":
          description: not found
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
  /api/animals/{id}/neutered:
    patch:
      tags: [Animals]
      summary: Update animal neutered boolean
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [neutered]
              properties:
                neutered: { type: boolean }
      responses:
        "200":
          description: updated
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Animal" }
        "422":
          description: validation failed
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
        "404":
          description: not found
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }

  ########################################################
  # Adoptions (mounted: /api/adoptions)  ※ 내부가 /adoptions...
  ########################################################
...